+++ rendering/PCDX11RenderModelInstance.h
@@ -22,7 +22,7 @@ public:
 	}
 
 	void resRegisterAtRenderDevice() override { /*TODO*/ };
-	void resFill(void* src, size_t size, size_t offset) override { /*empty*/ };
+	void resFill(void* src, uint32_t size, uint32_t offset) override { /*empty*/ };
 	char *resGetBuffer() override { return nullptr; };
 	void resConstruct() override { /*empty*/ };
 	void resMethod10() override { /*empty*/ };
